requirements.txt
fastapi
uvicorn
pypdf
google-genai
python-multipart